Initial code to add a third control track field to a profile.
A line like the curretn en,hhr becomes en,hhr,control in order to
use control.txt as the control track.

The code to parse the above line and fill in data structures is included.

NOTE: The third field is parsed but no code is here that actually uses it.