Better finding of the Dropbox location.
With this change we now use the Dropbox configuration to find the user's Dropbox installation.
Additionally, one can use the environment variable 'dropbox' to set up a temporary override,
useful for testing.

Dropbox is used as the "global shared directory", and is part of the ACMConfiguration. To simplify
setting up the configuration, it was converted from a collection of static properties to a
singleton shared instance; adding the 'getInstance()' calls is the bulk of the lines changed
(though the least interesting part).

The actual Dropbox finding code is added in utils/DropboxFinder.java, as well as unit tests for that
new code.

With junit and mockito/powermock integrated into the project, it will be easier for future
commits to also include unit tests. They will now be expected :)

Still remaining is to refactor the Ivy and build.xml code to not deploy