Introduce new interface called DataChangeListener; such a listener can be registered with MetdadataStore to receive notifications whenever an item was committed to the store due to an add, delete or modification event.
Change AudioItemTableModel to use the new DataChangeListener interface to notify the JTable of content changes incrementally, instead of always creating a new instance of AudioItemTableModel when the content set has changed. Now the model always contains all AudioItems, independent of the current search result. A RowFilter is instead used to hide AudioItems from the view that did not match any search results. These changes allow the JTable to sort columns more efficiently, or even to avoid resorting altogether.
Another change is to materialize the table cell values in AudioItemTableModel instead of recomputing them every time getValueAt() is called. This results in a meaningful speedup especially for values in the DATE_FILE_MODIFIED column; before for this column the .a18 files on disk has to be checked every time the table was rendered.
Lastly, this commit adds a specific Comparator for the PLAYLIST_ORDER column to perform a correct numeric sorting.

Cleanup code in AudioItemView and AudioItemTableModel that deals with columns by moving most code defining table columns to a new class called ColumnInfo.

Simplify AudioItem caching in LuceneMetadataStore; remove CommitListener in favor of the new DataChangeListener

Improve performance of the RowFilters used for the AudioItem table.

Addressed Bill's code review comments.