Set language on multiple audio items at once.
Previously, setting the language for multiple audio content items required editing
the items one by one. This allows selecting multiple audio items and setting the
language for all at once.

The importer always forces the language to English, which is usually wrong, because
none of the real content is in English.