[ACM-25] Implement a filtering function for communities in TB-Loader.
Some projects have many communities, and scrolling through the dropdown list to
find a particular one can be very time consuming and frustrating. This change
implements a "Community filter" that will limit the list of displayed communities
to those that contain a user entered string.

The filter is completely optional; if ignored, the app works exactly as before.