Implement Base 26 Rotating Encoding
This is like a base-26 encoding, except that it rotates the symbol list for each
more significant digit based on the symbol for the less significant digit.

The point is that the encoding for x and x+1 will be completely different, including
(or especially) in the more significant digits. This makes it easier to copy the
number without accidently jumping to the line below or above.

And although the strings look random-ish, they're completely deterministic and reversible.