Make locking work on OS/X as well as Windows.
The locking code that was to prevent multiple instances of the ACM application
running for a single content database had a bug on non-Windows platforms, due
to deleting the lock file. On Windows, the delete failed if another application
had the file open; on non-Windows systems, the file is deleted, and the other
application continues running on a shadow copy fo the file. The new instance then
opens the new file, and locks it.

Fix is to not delete the file prior to trying to lock it; only when cleaning up.