Fix bugs in TB-Loader.
If the SD card were blank or the 'system' directory were missing, the
lastSynchDir could not be read, and was left null, leading to a couple
of ignored NullPointerExceptions. The second of these wound up writing
only part of the line of log data, corrupting that line and the next
(which then provided the newline to complete the log record).