Address bug in TB-Loader gathering statistics.
The TB-Loader never set the value of the project anywhere in the Talking Book.
Thus, the Talking Book has NO IDEA which project it has been updated with. When
collecting statistics and user recordings the TB-Loader assumed that the new
project was the same as the old project. This is often true, because Talking
Books don't move between projects much. But when they do, the stats and feedback
are recorded to the wrong project.

This update addresses that by starting to record the current project into the
Talking Book. When reading statistics, the it will attempt to read the project
name from the TB. If the TB has been updated by this version of the TB-Loader or
later, it will have the project, and all's well. Otherwise, the TB-Loader will
fall back to the old behaviour of using the new project's name.