Fix bugs in community filter.
If a specific community is not selected ("Non-specific"), then the user is
prompted to confirm. Was predicated on the value "Non-specific" being at the
first index. Now checks by name.

When a new TB was pluggend in, attempting to fill the community list caused an
exception, which disabled the UI.