Hack that seems to help the reliability of writing to TB from Android.
After the update is done, walk the directory tree again. Possibly it is the
additional 10 seconds, or possibly it is the scanning of the files. It seems
to help. The resulting directory listing is written to a file in the Temp
directory, so that may help in troubleshooting, as well.