Improve .rev file calculations in TB-Builder.
Previously, we looked for DEPL-x.rev, and chose x+1 for the new rev (so, a->b, b->c, etc.).
But, if a previous deployment was made "current", next TB-Builder would overwrite then one
after it.

Now, we find the x.rev, then look for the next un-used letter. If no such letter, abort. This
means that we can't have more than 25 revisions to the original deployment.