Limit the length of a "Content Package" name.
The Talking Book limits the length of the Package name to 20 characters. The
ACM had no enforcement of this, and would happily produce a Package that could
not be played.

This change limits the length of the Package name to 20 charactes, and gives
a running count of the remaining length.