Much better test for TB-Loader.
This implements a test of the TB-Loader core update mechanism. The test:
- Creates a "virtual TB", by copying files taken from Talking Book prior
to an update. Rather than the actual live content, the files each
contain a small unique number. This makes the size quite small, and
makes testing for changes very easy.
- Updates that virtual TB with files from a "virtual Deployemnt". Like
the virtual TB, the files in the virtual Deployment have contents that
is a small unique number (also unique wrt the vTB).
- Examines the results of updating the vTB. Most files in a Deployment
are copied from the Deployment files to the TB, and those files are
checked that they were, in fact, copied, that all of the expected files
were copied, and that no unexpected files were copied. Of the non-copied
files, the most interesting is "deployment.properties", containing
metadata about the Deployment; the contents of that file are read, and
checked against expected values.

There are a variety of other tests that could be built:
- different update scenarios
- stats-only scenarios
- checking the files copied off of the vTB
- many more.