Significant improvement to category filtering.
It is now possible to blacklist and whitelist categories for display to
user in the ACM. (Only the categories that are shown can be assigned to
a message.) Also, visibility of parent categories is considered when
deciding to make a category visible or not, allowing a reasonably
compact specification.

Also performed some more refactoring on the Taxonomy loader to make it
more readable. It is now much more readable, and better isolated from
the Taxonomy using code.