Be less restrictive moving files for processing.
Prior to this change, we moved files into collected-data-processed if
they had {PROJ} subdirectories such that EVERY {PROJ} had a
TalkingBookData (TBD) subdirectory. A rogue directory without a TBD
subdir would prevent any files being copied and processed.

With this change, we copy files if ANY has a TBD subdir.

A side effect is that it allows more badly-formed data through to the
next step. But the next step already needs to be able to handle any kind
of bad, missing, or corrupted data, so this is not a problem.