Implement "User Feedback Hidden" support.
If the _activeLists.txt file that is used by TB-Builder to create a
deployemtn (actually a package withing a deployment) does not have a
"9-0" User Feedback category, then the package is generated as "User
Feedback Hidden". That involves using a different control.txt file, with
a "UFH" option in the header of the file, and a different message to be
played when user feedback recording is completed.

User feedback can still be recorded as always, but it is played back one
and only one time to the user. After that it is "hidden" until we
retreive user feedback for analysis.

The TB-Builder detects the lack of a "9-0" category and copies the
correct control.txt template.

The validation has been improved, to warn the user that the package will
have hidden feedback. The system prompts are now checked for presence
(because there is a newly required message that must be recorded for
existing languages). And the firmware version is now checked (because
the feature requires a firmware update on the device).