Translate language code "ssl1" => "sil" on db load.
A number of years ago (I write this on 2018-05-10), we needed a new language, Tumu Sisaala.
The person implementing the language did not know the ISO 639-3 code, nor did he know that
he should strictly restrict the codes to that well-known list. He just made up "ssl1", as
a modification of Lambussie Sisaala, "ssl". It wasn't a bad approach at the time, but it
was not correct, the correct code should have been "sil".

We've lived with this, as I say, for years. But today the pain of keeping the non-standard
language code exceeds the cost of fixing it, and so, here we are. This code translates
"ssl1" => "sil". It's only needed once per ACM, after which it adds perhaps 1ms to start
up time.