Finish Deployment Manager in TB-Loader.
The TB-Loader now manages deployments, automatically copying the latest
update when appropriate, prompting the user whether to test with an
unpublished version when appropriate.

If a project is configured with "DEPLOYMENT_CHOICE=TRUE" (or TB-Loader
is run with a "-c" option), the user is prompted to select from a list
of deployments. (CBCC needs this.)