Add UUID to deployment.properties.
With this change we assign a new UUID to every deployment on each
Talking Book. This will allow us to reliably correlate usage statistics
with an individual deployment, per Talking Book.

Also write the UUID to the deployments.log file and the tbData.log file,
and the previous UUID to the tbData.log file and the statsData.log file.