Implement "Recipient Chooser".
This introduces an improved TB-Loader (desktop & mobile) with a
"recipient chooser" to more easily navigate to a recipient, by
navigating through the administrative (or geo-political) heirarchy to
narrow down the set of recipients, ultimately to one.

The recipient chooser is a replacement for the community chooser, which
was simply a dropdown of all the community/groups in the project.

The old community chooser is temporarily still available, by adding a
line
    OLD_COMMUNITY_CHOOSER=TRUE
to the config.properties file of the project.

The implementation uses a new "HierarchicalList" which provides a
hierarchical view of a list in which certain fields (eg "country",
"region", or "district") are themselves the layers of a hierarchy. A
subclass, "DelayeredHierarchicalList" presents a flattened view, in
which the layers that provide no navigation are omitted from the view
(so, if all of the records are in the "Jirapa" district, then district
is omitted from the view, because there's no meaningful choice to be
made.)

There is also some fit-and-finish work on the desktop TB-Loader, so that
it is organized a bit more logically, and presented a bit more
attractively.