Include a 'programspec' directory with deployments.
Now keeps a copy the contents of the ACM-XYZ/programspec directory with
the deployment. The TB-Builder first copies the directory to the staging
area in a directory named
    ~/LiteracyBridge/TB-Loaders/{proj}/content/{deployment}/programspec

When published, it gets included into the deployment simply by virtue of
existing in the {deployment} directory. It is also copied into a
directory named
    ~/Dropbox/ACM-{proj}/TB-Loaders/{deployment}-x/programspec
so that it is available without needing to unzip anything.