Implement AcmCleaner.
The AcmCleaner tool will delete files from the ACM content directory
that have no matching entry in the (Lucene) database. This is the same
as using the --clean option when launching the ACM.

As part of the implementation, the database open code is enhanced to
return true if the database is successfully opened as requested.