Implement strict deployment naming in TB-Builder.
If there is a programspec directory, and it contains a deployments.csv
file, TB-Builder reads the file, and tries to match the specified
deployment with an entry in the file.

If no match is found, an error is given, and the operation does not
continue. An attempt is made to suggest what the desired deployment name
might have been.

To disable this, add the following line to config.properties for the
ACM:
    STRICT_DEPLOYMENT_NAMING=FALSE

This has been done for the testing and demonstration ACMs.