Use the language code as entered in the config.properties.
Previously we obtained the language code from Locale.getISO3Language(),
which had the effect of recording English as "eng". Everywhere else, we
use "en" for English, so this caused problems matching the audio to
English. Now we use Locale.getLanguage().

Or, as Bill Monroe said, "Play it like it was wrote, boys".