Added an ACM chooser dialog.
If there is no ACM given on the command line, we now look at the ACMs
on the user's machine, and let them choose from a list. If there is an
ACM on the command line, operation is unchanged.

This was inspired by the desire to support a beta version that would be
able to open whaterver ACM the user wished. It can also be useful to
reduce the desktop clutter of a dozen ACM icons.