Fix the annoying bug that clicking on a playlist didn't *quite* toggle the state.
The main window has a feature whereby all of the elements can be notified of changes
to the current search state. However, some of the elements adjust themselves
based on that notification, causing new notifications. Fortunately, the loop
sputters out after one iteration, but the damage to state has already been
done.

This change works around the architectural bug by not updating one of the
elements (the playlist list) unless the fundamental list has changed (ie,
a playlist added, removed, or renamed). The bug's still there, but losing
selections when the list structurally changes isn't bad, from a UX point
of view.