Enhance Dropbox Finder by looking for LB software.
Now look for presence of LB-software/ACM-install/ACM/software/acm.jar to
identify the proper Dropbox directory.

This will let us distinguish between an installation with both a
Business and a Personal dropbox, where the LB software is in the
Personal dropbox.

Still prefers Business, so if the .jar is found in Business, won't even
try to look in Personal.