Eliminate Dropbox for TB-Loader, implement server managed SRN allocation blocks.
The desktop application has been brought into alignment with the Android
application in using AWS S3 to distribute Deployments and to collect
usage statistics and user feedback. In order to access S3, the user
needs to be authenticated with an Amplio User ID.  That was already
implemented on Android and has now been implemented in the desktop
application. The Android implementation has been enhanced to use the
metadata returned in the sign-in call to avoid needing a separate query
for user attributes.

We keep track of "tb loader id" by user's email address. For each user,
we allocate blocks of SRNs. These blocks are persisted locally on the
device (laptop or Android device). When the SRNs are half used (ie, only
half of the original allocation remains), we'll allocate another block
when we can. If the same user uses the TB-Loader on multiple devices,
they'll get independent blocks on each device. These SRN blocks are
managed by a Lambda application.

Both of these features can be disabled by setting a flag in
config.properties:
    TBLOADER_DROPBOX=false

To implement the desktop sign-in:
- Added some code (from Amazon) to sign in via Cognito.
- Built a sign-in dialog, and a forgot password dialog.
- Persist the users' sign-in information for use when offline, and
  to pre-populate the sign-in dialog next time.

To implement the S3 access:
- Get a list of available Deployments (for the user) from S3.
- Download Deployments from S3, rather than copying from a local
  directory.
- Collect statistics and user feedback, and zip them into a single file
  on a per-TB basis, as the Android app does.
- When online, upload any collected files that haven't yet been
  uploaded.

Other changes:
- Updates to the build system for Android.
- Checked .idea files into git.
- Fix Android app to only show a user's projects in "Manage Content".
- Use apache.commons.lang3 everywhere (no more apache.commons.lang).

Unrelated change that got mixed in:
- Now support "invitation" and "action" as the decorator in the filename
  for a long Playlist prompt. Now easy to add further decorators.