Implement "Variant" in Program Specification.
The simplest description of Variant is that it is useful to implement A/B testing.
A Recipient can be marked with a Variant (free form text).
Messages can be marked with one or more Variant(s).
A message with no variants goes to all Recipients. Messages with a variant only go
to the Repients that match the variant.

The implementation involved:
- Adding an optional Variant column to Recipient and Content tables.
- Adding the parsers and accessors for Variant.
- Potentially creating multiple "Images" per language in a Deployment.
- Associating Recipients with the proper Image.
-- This is computed in the "Create Deployment" phase, and persisted in a
   deployment.properties file that is saved along with the program spec
   for a deployment.
- In the TB-Loader, putting the correct Image onto a Talking Book.
-- Based on the Image computed when the Deployment was created.