In TB-Loader, don't reset the selected recipient so eagerly.
When a TB is connected, the TB-Loader pre-populates the recipient
chooser from the recipient of the connected TB. If the recipient is not
from the current program, the values are cleared. This makes loading new
TBs more difficult, especially when they're all for the same community.

With this change, only reset the recipient if it matches an existing
recipient in the program.