Fix a problem causing duplicate ids when importing audio content.
For a reason that is still unidentified, LBG wound up with 3 instances
of multiple audio items with the same id. In any normal operation, this
should not happen. To occur:
- The audio file being imported must be an .a18 file with an id in its
  embedded metadata. The id is of an existing item in the same ACM.
- The audio file is imported as a new item, and gets a new, different
  file name.

There is a defined mechanism to import a new item with the same id,
which is to append "___" and the id to the message title.