Fix a bug with forced password resets.
When the admin resets the password, the user must change it next time
they sign in. If they have never signed in, we may not know their email
address (which is not currently what we sign in with). When we get to
the "reset" panel, we were requesting the email, and the password change
failed.

Now use email if we have it, otherwise user name.