Convert the default L&F from SeaGlass to FlatLaf.
FlatLaf is a new, and supported, L&F. See https://www.formdev.com/flatlaf/

- Some tweaks to UIManager defaults with FlatLaf, to make ACM  work with FlatLaf.
- Small fix to disable the "Desktop Shortcuts" panel when there are no
  desktop shortcuts.
- Added a --devo option to open a UIManager properties dialog.
-- Intent is to be able to see the properties to get from the UIManager,
   rather than hard coding colors in dialogs and buttons.
- Tweaks to make drag-and-drop look OK with darklaf.
- Use Lucida Grande as the default font with FlatLaf. This is the same font
  that SeaGlass used.
- Use the latest Amplio logo in the sign-in dialog.
-- Also factor out the code that sets the logo in every panel.
- UI tweaks to make FlatLAF look nicer.

With this, FlatLAF looks pretty good, and is a suitable replacement
for SeaGlass.  I still think SeaGlass is the most attractive of the
pluggable LAFs. But it is no longer supported, and doesn't work on
Java past Java 8, so we need a replacement to keep up with Java.