Honor "deployments" column in Recipients list.
A "deployments" property has been added to Recipients, which is a list of the
deployments that a recipient should get. It is a per-recipient list, in JSON
format (sigh, a comma separated list would have been easier in every way). If
empty or missing, the recipient should get all deployments, but if there is a
list, only those in the list.