Improvements to audio player.
Fixes to media player.
- Fixed some race conditions that could lead to two audio tracks playing at
  once.
- More rational API to simple player.
- Commented the public API.

Improvements to the media controls.
- Added a "Stop" button.
- Can now click "Play" to play highlighted track.
- Scrubbing now updates the display appropriately.
- Scrubbing now works when paused.
- Took (incorrect) pixel count out of icon filenames.

Fix media player title so that it doesn't grow.

The media player title used to grow to be as big as it wanted. This would
push the search box off of the screen. Now restrict it to the size of the
scrubber.

It might be nice ot add a marquee display, to show the entire title.