Delete seaglasslookandfeel from 3rd party libraries.
SeaGlass has been replaced with FlatLaf. Although SeaGlass is one of the
nicest, if not the nicest, of the look-and-feel libraries, it is no
longer supported, and does not work on Java versions newer than Java 8.
This has prevented us from moving to a newer version of Java.

FlatLaf is a modern, supported look-and-feel. It includes niceties such
as a github repository with full source, and theme support.