Implemented a general "sandbox" facility. Will improve & augment content sandboxing.
- The "sandboxer" is initialized with the directory to be sandboxed (or
  shadowed) and a directory to hold the interim changes.
- An API allows:
-- testing for existance
-- opening files for read or write (as Streams)
-- renaming files within the sandboxed directory
-- moving files into the sandbox from outside
-- committing the changes from the sandbox to the "real" (base)
   directory.

When changes are committed, callbacks allow handlers to monitor the
actual changes. This will be used to schedule updates to S3 when content
or configurations change.

This is a superset of the content sandbox facility, although that is
built specifically for content and content only.  With this it will be a
simple matter to sandbox other files, such as configuration files or
deployment files.