Added a hint to help builders sign AmplioSetup more reliably.
The signing tool can appear to work even if the token is not connected,
but the signature will not be an "Enhanced Validation" signature, and
the end user will still be warned that the setup will melt their
computer.

This change reminds the user that unless they're prompted to enter the
token password, proper signing will not take place, all appearances to
the contrary notwithstanding.