Capture TBv2 stats data in a simpler format.
The TBv2 statistics are saved into an incredibly deeply nested directory
hierarchy (seriously, for example: CARE-BGD-JANO/TalkingBookData/
CARE-BGD-JANO-22-1/0070/DOMAR-MEGHNA_JAMUNA_KARNAFULI-REKHA_AKTER-BEN/
C-00700C39). Metadata about the statistics or the deployment were coded into
the various directory names. For quite some time, we've included a file with
the metadata, so the directory names are redundant, not to mention extremely
difficult to work with. However, to avoid having to change the stats parsing
application, we've preserved the directory structure.

The new TBv2 stats are arriving in a different format, and the parsing will
necessarily be different and new. There is no need to keep the complicated
directory structure. This change abstracts out the creation of the paths into
which the stats are copied, and generates appropriate paths for v1 and v2.

This also refactors the creation of the stats logs out of the "TbLoaderCore"
file. It is still way too long, but now "only" ~1200 lines.