Fix for updated program spec in deployments.
There was a bug that when the program spec was updated, the deployment
was created with the old files, from the program directory, not the new
files sitting in the sandbox directory.

With this change, the tb builder uses Sandbox to find the current file, and
copies the current program spec files. Also copies only program spec files,
and not any other cruft sitting in the directory.