Fixes to display info about TB from another program.
If a Talking Book was last used in a different program, the recipientid
will not be meaningful in the new program, and so the "previous
recipient" won't be able to fill.

This change uses information from the deployment.properties on the
Talking Book to show where it was before. It also adds more data to the
deployment.properties file, so that this will be better in the future.