Fixes to LuceneMetadataStore transaction handling.
The Import Assistant creates and populates playlists based on the program
specification. Due to a bug in how batch update transactions are handled,
when multiple playlists are added in a single transaction, only the last
one "sticks"; each "new playlist" update overwrites the previous one.

This is due to how playlists are handled. The list of playlists is kept as an
extra bit of data (it is not clear to me what the Lucene terminology is), and
the "new playlist" update overwrites that data with a new list, built from the
currently committed list plus the new playlist. This change updates the list
with a new list built from the un-committed list plus the new playlist.

This only works at all because the only thing that this extra bit of storage
is used for is playlists.

The proper solution would be to migrate from Lucene to an actual database.