Fix to detect all UF recordings on Talking Book.
User Feedback recordings on the Talking Book are recognized by the TB-Loader
based on the file name. The recording is named with the TB SRN as a prefix,
followed by either a category ("9" or "9-0") or content id (like
"LB-2_6hyolivfgp_kb").

Previously, the TB-Loader recognized only the "9" and "9-0" forms of UF
files. This is a very long-standing bug, a bit over 5 years old. It hasn't
mattered until now because all actual UF recordings are of the "9-0" form.
However, the custom surveys for Berhan created the "contentid" form, which
we then failed to collect.

- Now recognize all UF files, by detecting SRN at the start of file name.