Fix an ancient bug in playlist management.
Playlists are stored somewhat weirdly, to put it kindly. There is an
"additional data" string that is used to store playlists for an audio
item. The string consists of the audio item id, playlist ids, and
playlist names concatenated into a string. The delimiter characters that
were used were "," and ":". With no checks to prevent the use of those
characters in a playlist name, it was possible to create a playlist that
would be parsed incorrectly.

The unfortunate side-effect of this is that the ACM could not be re-opened
without manual manipulation of the database. The good news is that, since
this hasn't happened (yet), nobody has used a comma in a playlist.

The solution here is to use some obscure Unicode symbols for the concatenation
characters (\u22a1 and \u2299, "square with dot" and "circle with dot", from
the mathematical symbols section). Further, if those characters are found
in a playlist's name, they're replaced with space.

It is not impossible that someone has used one of those symbols in a playlist
name already, but the odds are vanishingly small.