Fix a bug loading TB-Loader history.
If an ACM database hadn't been loaded by the ACM, the TB-Loader would NPE
trying to open it. This was due to the program directory not existing in the
'Amplio/acm-dbs' directory, which prevented the PathsProvider from being
pre-created. Now we create the PathsProvider object on demand, whether or not
there is program data on the computer.