Fixes for tightened Windows security.
On some Windows systems, attempting 'chkdsk D:' of the Talking Book
device fails with access denied. Detect that, try to update anyway,
and warn the user that they may need to run chkdsk /f manually.

Also removed the code to disconnect the drive. That hasn't worked for
quite a while, and is also not necessary on Windows (it seems they
don't cache USB writes).