Feature: Generate .csm scripts for custom surveys in deployment.
For playlists marked as survey (currently by means of the short prompt
having the category "Survey"), with a corresponding playlist.survey file
in the program spec directory, the .survey file is compiled into a
custom .csm script to implement the survey.

Because the ACM now uses CSMcompile when creating a survey CSM script,
it is now imported from GitHub.