Fix: When updating a TBv1, prefer 'images.v1' over 'images'.
This is the next step in a change started long ago. With the advent of TBv2,
deployments need images for both v1 and v2 Talking Books. The TB-Builder
has been creating both 'images' and 'images.v1' but the TB-Loader continued
to look only for 'images'. With this change it now looks for 'images.v1'
first and only if that does not exist will it look for 'images'. Once this
has rolled out everywhere, we can stop creating the 'images' version, and
can remove the code from the TB-Builder that creates it.