refactor: archived 'androidTBloader' app project. Renamed dir to `archivedAndroidtbloader` and namespace to `org.literacybridge.archived_androidtbloader`
The new TB loader app for TBv2 & 1 uses jetpack compose + major structural differences
due to android studio changes. Although we want to publish the new app under the same package name,
we also have to maintain the codebase of the old app, should we have to revisit for reference.

To achieve this, I archived (renamed project to 'archivedAndroidtbloader') and changed the namespace
from 'org.literacybridge.androidtbloader' to 'org.literacybridge.archived_androidtbloader'.

In the next commits, namespace of the new project will be changed to 'org.literacybridge.androidtbloader',
in other to publish it as an update on PlayStore instead of as a new app.

(cherry picked from commit a97145cbe4dd3833e500c899322d7dac7504f116)